+++ Makefile
@@ -30,7 +30,8 @@ tests:
 .PHONY:lint
 lint:
 	. ${NVM_DIR}/nvm.sh && nvm use && \
-	npm run lint
+	npm run lint && \
+	npm run typecheck
 
 .PHONY:fix-lint
 fix-lint:
