+++ src/registry/extractor_registry.ts
@@ -1,4 +1,4 @@
-import { BaseResourceExtractor, Extractor } from './base_extractor';
+import { BaseResourceExtractor, Extractor } from '../converters/base_extractor';
 import { TResource } from '../types/resources/Resource';
 
 export class ExtractorRegistry {
