+++ test/converters/fhir_extractors.spec.ts
@@ -1,10 +1,10 @@
 import { PatientExtractor } from '../../src/converters/patient_extractor';
 import { ObservationExtractor } from '../../src/converters/observation_extractor';
-import { ExtractorRegistry } from '../../src/converters/extractor_registry';
-import { FHIRBundleConverter } from '../../src/converters/fhir_bundle_converter';
+import { ExtractorRegistry } from '../../src/registry/extractor_registry';
+import { FHIRBundleConverter } from '../../src/fhir_bundle_converter';
 import { TBundle } from '../../src/types/resources/Bundle';
 import { TPatient } from '../../src/types/resources/Patient';
-import { ExtractorRegistrar } from '../../src/register';
+import { ExtractorRegistrar } from '../../src/registry/register';
 import { TResource } from '../../src/types/resources/Resource';
 import * as fs from 'node:fs';
